requirement.txt
pymysql>=1.0.2
pyjwt>=2.3.0
flask>=2.0.3
flask_sqlalchemy>=2.5.1
cryptography>=36.0.1
